Extract `close_trusted_market` tests
